requirements.txt
dataclasses==0.7;python_version<3.7
elasticsearch==7.0.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
google-cloud-storage==1.19.1
